Moves away from eval.go to store_eval.go and from async server to resp
